`helpers`: move `term.lua` up from `lua_ext/`.
