+++ mods/lord/Core/helpers/init.lua
@@ -1,9 +1,9 @@
 local mod_path    = minetest.get_modpath(minetest.get_current_modname())
 local old_require = require
-require           = function(name) return dofile(mod_path .. "/src/" .. name:gsub("%.", "/") .. ".lua") end
+require           = function(name) return dofile(mod_path .. '/src/' .. name:gsub('%.', '/') .. '.lua') end
 
 
-require("lua_ext")
+require('helpers')
 
 
 require = old_require
