+++ mods/lord/Core/helpers/src/helpers.lua
@@ -1,9 +1,8 @@
-
+require('term')
 require('lua_ext.global')
 require('lua_ext.table')
 require('lua_ext.string')
 require('lua_ext.math')
 require('lua_ext.io')
 require('lua_ext.os')
-require('lua_ext.term')
 require('lua_ext.debug')
